chore: Better files org
