+++ src/GmdVizPanel/components/ConfigurePanelAction.tsx
