+++ src/GmdVizPanel/components/EventConfigurePanel.ts
