+++ src/GmdVizPanel/config/presets/getConfigPresetsForMetric.ts
@@ -1,13 +1,12 @@
-import { isHistogramMetric } from 'GmdVizPanel/heatmap/isHistogramMetric';
-import { isAgeMetric } from 'GmdVizPanel/isAgeMetric';
-import { isRateQuery } from 'GmdVizPanel/isRateQuery';
-import { isUpDownMetric } from 'GmdVizPanel/statushistory/isUpDownMetric';
-
 import { DEFAULT_TIMESERIES_AGE_PRESETS } from './config-presets-ages';
 import { DEFAULT_HISTOGRAMS_PRESETS } from './config-presets-histograms';
 import { DEFAULT_STATUS_UP_DOWN_PRESETS } from './config-presets-status-updown';
 import { DEFAULT_TIMESERIES_PRESETS, DEFAULT_TIMESERIES_RATE_PRESETS } from './config-presets-timeseries';
 import { type PanelConfigPreset } from './types';
+import { isCounterMetric } from '../..//matchers/isCounterMetric';
+import { isHistogramMetric } from '../..//matchers/isHistogramMetric';
+import { isUpDownMetric } from '../..//matchers/isUpDownMetric';
+import { isAgeMetric } from '../../matchers/isAgeMetric';
 
 export function getConfigPresetsForMetric(metric: string, isNativeHistogram: boolean): PanelConfigPreset[] {
   if (isUpDownMetric(metric)) {
@@ -22,7 +21,7 @@ export function getConfigPresetsForMetric(metric: string, isNativeHistogram: boo
     return [Object.values(DEFAULT_TIMESERIES_PRESETS)[0], ...Object.values(DEFAULT_TIMESERIES_AGE_PRESETS)];
   }
 
-  if (isRateQuery(metric)) {
+  if (isCounterMetric(metric)) {
     return Object.values(DEFAULT_TIMESERIES_RATE_PRESETS);
   }
