+++ src/GmdVizPanel/matchers/isAgeMetric.ts
@@ -1,4 +1,4 @@
 /**
  * Identifies gauge metrics with a value measured as a Unix timestamp in seconds
  */
-export const isAgeMetric = (metric: string) => metric.includes('timestamp_seconds');
+export const isAgeMetric = (metric: string) => metric.endsWith('_timestamp_seconds');
