+++ src/GmdVizPanel/matchers/isHistogramMetric.ts
