+++ src/GmdVizPanel/matchers/isUpDownMetric.ts
