+++ src/GmdVizPanel/types/statushistory/value-mappings.ts
